Refactor Admin Layout
