+++ apps/web/src/components/pages/admin/manage-event/promotions-codes.tsx
@@ -1,4 +1,4 @@
-import { TropTixContext } from '@/components/WebNavigator';
+import { TropTixContext } from '@/components/AuthProvider';
 import { Spinner } from '@/components/ui/spinner';
 import { Promotion, createPromotion } from '@/hooks/types/Promotion';
 import {
