+++ apps/web/src/hooks/useFetchEvents.tsx
@@ -1,4 +1,4 @@
-import { TropTixContext } from '@/components/WebNavigator';
+import { TropTixContext } from '@/components/AuthProvider';
 import { useQuery } from '@tanstack/react-query';
 import { useContext } from 'react';
