+++ apps/web/src/pages/order-confirmation/index.tsx
@@ -1,4 +1,4 @@
-import { TropTixContext } from '@/components/WebNavigator';
+import { TropTixContext } from '@/components/AuthProvider';
 import { Spinner } from '@/components/ui/spinner';
 import { useFetchOrderById } from '@/hooks/useOrders';
 import { getDateFormatter, getFormattedCurrency } from '@/lib/utils';
